warning fixed

+++ mode.cc
@@ -16,10 +16,10 @@ std::vector<Track *> ShuffleMode::re_arrange(const std::vector<Track *> &v) {
 
 std::vector<Track *> OddEvenMode::re_arrange(const std::vector<Track *> &v) {
     std::vector <Track*> ordered;
-    for (int i = 1; i < v.size(); i += 2) {
+    for (size_t i = 1; i < v.size(); i += 2) {
         ordered.push_back(v[i]);
     }
-    for (int i = 0; i < v.size(); i += 2) {
+    for (size_t i = 0; i < v.size(); i += 2) {
         ordered.push_back(v[i]);
     }
     return ordered;
